requirements.txt
aniso8601==7.0.0
asgiref==3.4.0
dj-database-url==0.5.0
Django==3.2.4
psycopg2-binary==2.9.1
django-heroku==0.0.0
django-tinymce==3.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
Pillow==8.2.0
promise==2.3
pytz==2021.1
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
whitenoise==5.3.0
